refactor: move non-protocol related code to domain
- This structures the project code after well known practices where the
protocol is the rules and conventions that govern how data is
transmitted and communicated over the network, and the domain refers to
the subject area and problem space that the softweare is designed to
address. It represents the business logic, processes, and rules
associated with the specific features and services of the system.
