+++ safenode/src/client/register/mod.rs
@@ -12,7 +12,7 @@ pub use offline_replica::RegisterOffline;
 
 use super::{error::Result, Client};
 
-use crate::protocol::storage::register::{Entry, EntryHash, Policy};
+use crate::domain::storage::register::{Entry, EntryHash, Policy};
 
 use std::collections::BTreeSet;
 use xor_name::XorName;
