+++ safenode/src/domain/client_transfers/error.rs
