+++ safenode/src/domain/client_transfers/offline.rs
