+++ safenode/src/domain/fees/error.rs
