+++ safenode/src/domain/fees/required_fee.rs
