+++ safenode/src/domain/fees/required_fee_content.rs
