+++ safenode/src/domain/node_transfers/error.rs
@@ -6,7 +6,7 @@
 // KIND, either express or implied. Please review the Licences for the specific language governing
 // permissions and limitations relating to use of the SAFE Network Software.
 
-use crate::protocol::{fees, storage::DbcAddress};
+use crate::domain::{fees, storage::DbcAddress};
 
 use sn_dbc::{Error as DbcError, SignedSpend, Token};
