+++ safenode/src/domain/storage/address/register.rs
