+++ safenode/src/domain/storage/error.rs
