+++ safenode/src/domain/storage/mod.rs
