+++ safenode/src/domain/storage/registers/mod.rs
