+++ safenode/src/domain/storage/registers/register/authority.rs
