+++ safenode/src/domain/storage/registers/register/metadata.rs
