+++ safenode/src/domain/storage/spends.rs
@@ -6,7 +6,7 @@
 // KIND, either express or implied. Please review the Licences for the specific language governing
 // permissions and limitations relating to use of the SAFE Network Software.
 
-use crate::protocol::{
+use crate::domain::{
     node_transfers::{Error, Result},
     storage::DbcAddress,
 };
