+++ safenode/src/domain/wallet/keys.rs
