+++ safenode/src/domain/wallet/mod.rs
