+++ safenode/src/domain/wallet/network_store.rs
