+++ safenode/src/domain/wallet/wallet_file.rs
