+++ safenode/src/lib.rs
@@ -44,6 +44,8 @@ extern crate tracing;
 
 /// SAFE Client
 pub mod client;
+/// SAFE Domain Logic.
+pub mod domain;
 /// Logging.
 pub mod log;
 /// The main logic of the network.
