+++ safenode/src/protocol/error.rs
@@ -6,7 +6,7 @@
 // KIND, either express or implied. Please review the Licences for the specific language governing
 // permissions and limitations relating to use of the SAFE Network Software.
 
-use super::{node_transfers::Error as TransferError, storage::Error as StorageError};
+use crate::domain::{node_transfers::Error as TransferError, storage::Error as StorageError};
 
 use serde::{Deserialize, Serialize};
 use std::{fmt::Debug, result};
