+++ safenode/src/protocol/messages/spend.rs
@@ -6,7 +6,7 @@
 // KIND, either express or implied. Please review the Licences for the specific language governing
 // permissions and limitations relating to use of the SAFE Network Software.
 
-use crate::protocol::{
+use crate::domain::{
     fees::SpendPriority,
     storage::{dbc_address, DbcAddress},
 };
